move OSR setting to JB IDE
