+++ extensions/intellij/src/main/kotlin/com/github/continuedev/continueintellijextension/services/ContinueExtensionSettingsService.kt
@@ -2,6 +2,7 @@ package com.github.continuedev.continueintellijextension.services
 
 import com.github.continuedev.continueintellijextension.constants.getConfigJsPath
 import com.github.continuedev.continueintellijextension.constants.getConfigJsonPath
+import com.intellij.openapi.application.ApplicationInfo
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.*
 import com.intellij.openapi.options.Configurable
@@ -28,6 +29,7 @@ class ContinueSettingsComponent : DumbAware {
     val userToken: JTextField = JTextField()
     val enableTabAutocomplete: JCheckBox = JCheckBox("Enable Tab Autocomplete")
     val enableContinueTeamsBeta: JCheckBox = JCheckBox("Enable Continue for Teams Beta")
+    val enableOSR: JCheckBox = JCheckBox("Enable Off-Screen Rendering")
     val displayEditorTooltip: JCheckBox = JCheckBox("Display Editor Tooltip")
 
     init {
@@ -55,6 +57,8 @@ class ContinueSettingsComponent : DumbAware {
         panel.add(enableTabAutocomplete, constraints)
         constraints.gridy++
         panel.add(enableContinueTeamsBeta, constraints)
+        constraints.gridy++
+        panel.add(enableOSR, constraints)
         constraints.gridy++
         panel.add(displayEditorTooltip, constraints)
         constraints.gridy++
@@ -87,6 +91,7 @@ open class ContinueExtensionSettings : PersistentStateComponent<ContinueExtensio
         var enableTabAutocomplete: Boolean = true
         var ghAuthToken: String? = null
         var enableContinueTeamsBeta: Boolean = false
+        var enableOSR: Boolean = shouldRenderOffScreen()
         var displayEditorTooltip: Boolean = true
     }
 
@@ -106,6 +111,7 @@ open class ContinueExtensionSettings : PersistentStateComponent<ContinueExtensio
         val instance: ContinueExtensionSettings
             get() = ServiceManager.getService(ContinueExtensionSettings::class.java)
     }
+
 
     // Sync remote config from server
     private fun syncRemoteConfig() {
@@ -198,6 +204,7 @@ class ContinueExtensionConfigurable : Configurable {
                     mySettingsComponent?.userToken?.text != settings.continueState.userToken ||
                     mySettingsComponent?.enableTabAutocomplete?.isSelected != settings.continueState.enableTabAutocomplete ||
                     mySettingsComponent?.enableContinueTeamsBeta?.isSelected != settings.continueState.enableContinueTeamsBeta ||
+                    mySettingsComponent?.enableOSR?.isSelected != settings.continueState.enableOSR ||
                     mySettingsComponent?.displayEditorTooltip?.isSelected != settings.continueState.displayEditorTooltip
         return modified
     }
@@ -210,6 +217,7 @@ class ContinueExtensionConfigurable : Configurable {
         settings.continueState.enableTabAutocomplete = mySettingsComponent?.enableTabAutocomplete?.isSelected ?: false
         settings.continueState.enableContinueTeamsBeta =
             mySettingsComponent?.enableContinueTeamsBeta?.isSelected ?: false
+        settings.continueState.enableOSR = mySettingsComponent?.enableOSR?.isSelected ?: true
         settings.continueState.displayEditorTooltip = mySettingsComponent?.displayEditorTooltip?.isSelected ?: true
 
         ApplicationManager.getApplication().messageBus.syncPublisher(SettingsListener.TOPIC)
@@ -224,6 +232,7 @@ class ContinueExtensionConfigurable : Configurable {
         mySettingsComponent?.userToken?.text = settings.continueState.userToken
         mySettingsComponent?.enableTabAutocomplete?.isSelected = settings.continueState.enableTabAutocomplete
         mySettingsComponent?.enableContinueTeamsBeta?.isSelected = settings.continueState.enableContinueTeamsBeta
+        mySettingsComponent?.enableOSR?.isSelected = settings.continueState.enableOSR
         mySettingsComponent?.displayEditorTooltip?.isSelected = settings.continueState.displayEditorTooltip
 
         ContinueExtensionSettings.instance.addRemoteSyncJob()
@@ -236,4 +245,33 @@ class ContinueExtensionConfigurable : Configurable {
     override fun getDisplayName(): String {
         return "Continue Extension Settings"
     }
+}
+
+/**
+ * This function checks if off-screen rendering (OSR) should be used.
+ *
+ * If ui.useOSR is set in config.json, that value is used.
+ *
+ * Otherwise, we check if the pluginSinceBuild is greater than or equal to 233, which corresponds
+ * to IntelliJ platform version 2023.3 and later.
+ *
+ * Setting `setOffScreenRendering` to `false` causes a number of issues such as a white screen flash when loading
+ * the GUI and the inability to set `cursor: pointer`. However, setting `setOffScreenRendering` to `true` on
+ * platform versions prior to 2023.3.4 causes larger issues such as an inability to type input for certain languages,
+ * e.g. Korean.
+ *
+ * References:
+ * 1. https://youtrack.jetbrains.com/issue/IDEA-347828/JCEF-white-flash-when-tool-window-show#focus=Comments-27-9334070.0-0
+ *    This issue mentions that white screen flash problems were resolved in platformVersion 2023.3.4.
+ * 2. https://www.jetbrains.com/idea/download/other.html
+ *    This documentation shows mappings from platformVersion to branchNumber.
+ *
+ * We use the branchNumber (e.g., 233) instead of the full version number (e.g., 2023.3.4) because
+ * it's a simple integer without dot notation, making it easier to compare.
+ */
+private fun shouldRenderOffScreen(): Boolean {
+    val minBuildNumber = 233
+    val applicationInfo = ApplicationInfo.getInstance()
+    val currentBuildNumber = applicationInfo.build.baselineVersion
+    return currentBuildNumber >= minBuildNumber
 }
