+++ extensions/intellij/src/main/kotlin/com/github/continuedev/continueintellijextension/toolWindow/ContinueBrowser.kt
@@ -5,6 +5,7 @@ import com.github.continuedev.continueintellijextension.constants.MessageTypes
 import com.github.continuedev.continueintellijextension.constants.getConfigJsonPath
 import com.github.continuedev.continueintellijextension.`continue`.*
 import com.github.continuedev.continueintellijextension.factories.CustomSchemeHandlerFactory
+import com.github.continuedev.continueintellijextension.services.ContinueExtensionSettings
 import com.github.continuedev.continueintellijextension.services.ContinuePluginService
 import com.google.gson.Gson
 import com.google.gson.JsonObject
@@ -78,7 +79,8 @@ class ContinueBrowser(val project: Project, url: String) {
     val browser: JBCefBrowser
 
     init {
-        this.browser = JBCefBrowser.createBuilder().setOffScreenRendering(shouldRenderOffScreen()).build()
+        val enableOSR = ServiceManager.getService(ContinueExtensionSettings::class.java).continueState.enableOSR
+        this.browser = JBCefBrowser.createBuilder().setOffScreenRendering(enableOSR).build()
 
 
         browser.jbCefClient.setProperty(
@@ -274,44 +276,4 @@ class ContinueBrowser(val project: Project, url: String) {
         return """window.postMessage($jsonData, "*");"""
     }
 
-    /**
-     * This function checks if off-screen rendering (OSR) should be used.
-     * 
-     * If ui.useOSR is set in config.json, that value is used.
-     * 
-     * Otherwise, we check if the pluginSinceBuild is greater than or equal to 233, which corresponds
-     * to IntelliJ platform version 2023.3 and later.
-     *
-     * Setting `setOffScreenRendering` to `false` causes a number of issues such as a white screen flash when loading
-     * the GUI and the inability to set `cursor: pointer`. However, setting `setOffScreenRendering` to `true` on
-     * platform versions prior to 2023.3.4 causes larger issues such as an inability to type input for certain languages,
-     * e.g. Korean.
-     *
-     * References:
-     * 1. https://youtrack.jetbrains.com/issue/IDEA-347828/JCEF-white-flash-when-tool-window-show#focus=Comments-27-9334070.0-0
-     *    This issue mentions that white screen flash problems were resolved in platformVersion 2023.3.4.
-     * 2. https://www.jetbrains.com/idea/download/other.html
-     *    This documentation shows mappings from platformVersion to branchNumber.
-     *
-     * We use the branchNumber (e.g., 233) instead of the full version number (e.g., 2023.3.4) because
-     * it's a simple integer without dot notation, making it easier to compare.
-     */
-    private fun shouldRenderOffScreen(): Boolean {
-        val continuePluginService = ServiceManager.getService(
-            project,
-            ContinuePluginService::class.java
-        )
-        
-        // Check if useOSR is explicitly set in config
-        val useOSR = continuePluginService.coreMessenger?.config?.ui?.useOSR
-        if (useOSR != null) {
-            return useOSR
-        }
-
-        // Fall back to version check
-        val minBuildNumber = 233
-        val applicationInfo = ApplicationInfo.getInstance()
-        val currentBuildNumber = applicationInfo.build.baselineVersion
-        return currentBuildNumber >= minBuildNumber
-    }
 }
